databene-commons-core/src/main/java/org/databene/commons/converter/ParseFormatConverter.java
/*
 * (c) Copyright 2007-2012 by Volker Bergmann. All rights reserved.
 *
 * Redistribution and use in source and binary forms, with or without
 * modification, is permitted under the terms of the
 * GNU General Public License.
 *
 * For redistributing this software or a derivative work under a license other
 * than the GPL-compatible Free Software License as defined by the Free
 * Software Foundation or approved by OSI, you must first obtain a commercial
 * license to this software product from Volker Bergmann.
 *
 * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
 * WITHOUT A WARRANTY OF ANY KIND. ALL EXPRESS OR IMPLIED CONDITIONS,
 * REPRESENTATIONS AND WARRANTIES, INCLUDING ANY IMPLIED WARRANTY OF
 * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE OR NON-INFRINGEMENT, ARE
 * HEREBY EXCLUDED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
 * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
 * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
 * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
 * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
 * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
 * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 * POSSIBILITY OF SUCH DAMAGE.
 */

package org.databene.commons.converter;

import org.databene.commons.ConversionException;

import java.text.Format;
import java.text.ParseException;

/**
 * Converts a String to an object by using a java.lang.Format object's format() method.<br/>
 * <br/>
 * Created: 30.08.2006 19:48:09
 * @since 0.1
 * @author Volker Bergmann
 */
public class ParseFormatConverter<T> extends FormatBasedConverter<String, T> {

    public ParseFormatConverter(Class<T> targetType, Format format, boolean threadSafe) {
        super(String.class, targetType, format, threadSafe);
    }

    /** Converts an object to a String by using the format's format() method. */
    @Override
	@SuppressWarnings("unchecked")
	public T convert(String source) throws ConversionException {
        if (source == null)
            return null;
        try {
            return (T) format.parseObject(source);
        } catch (ParseException e) {
            throw new ConversionException(e);
        }
    }

}
